+++ frontend/components/comments/comment_show.jsx
@@ -3,7 +3,12 @@ import React from 'react';
 class CommentShow extends React.Component {
   destroySelf(e) {
     e.preventDefault();
-    this.props.destroyComment(this.props.comment.id);
+    let answer = confirm(
+      "Are you sure? This action can't be undone."
+    );
+    if (answer) {
+      this.props.destroyComment(this.props.comment.id);
+    }
   }
 
   renderDeleteButton() {
@@ -22,17 +27,22 @@ class CommentShow extends React.Component {
   render() {
     return(
       <section className="single-comment">
-        <section className="voting-body">
-          <section className="voting">
-            <button className="downvote"><i className="fa fa-thumbs-down" aria-hidden="true"></i></button>
-            <span className="score">12</span>
-            <button className="upvote"><i className="fa fa-thumbs-up" aria-hidden="true"></i></button>
+        <section className="single-comment-content">
+          <section className="voting-body">
+            <section className="voting">
+              <button className="downvote"><i className="fa fa-thumbs-down" aria-hidden="true"></i></button>
+              <span className="score">12</span>
+              <button className="upvote"><i className="fa fa-thumbs-up" aria-hidden="true"></i></button>
+            </section>
+            <section className="comment-body">
+              {this.props.comment.body}
+            </section>
           </section>
-          <section className="comment-body">
-            {this.props.comment.body}
-          </section>
+          {this.renderDeleteButton()}
         </section>
-        {this.renderDeleteButton()}
+        <section className="author">
+          {this.props.comment.user}
+        </section>
       </section>
     );
   }
